AI/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
